removed previous movie test to retry-AM
