--- src/test/java/com/springboothackathon/controllers/MovieControllerTest.java
@@ -1,48 +1,0 @@
-package com.springboothackathon.controllers;
-
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.web.servlet.MockMvc;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-public class MovieControllerTest {
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    void callsMovieAPI() throws Exception {
-        mockMvc.perform(get("/movies"))
-                .andDo(print())
-                .andExpect(status().isOk());
-
-    }
-}
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class MovieControllerWithMockedWebServer{
-
-    @LocalServerPort
-    private int port;
-
-    @Autowired
-    private TestRestTemplate restTemplate;
-
-    @Test
-    void callsMovieAPI(){
-
-        String result = restTemplate.getForObject("http://localhost:" + port + "/movies", String.class);
-
-        assertEquals(result, "Greetings from Spring Boot!");
-    }
-}
