--- src/test/java/cucumbertests/MovieControllerTest.java
@@ -1,12 +1,0 @@
-package cucumbertests;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
-import org.junit.runner.RunWith;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(features = "src/test/resources/features/movies",
-    glue = "cucumbertests.stepdefs")
-public class MovieControllerTest {
-
-}
