+++ src/test/java/cucumbertests/SpringIntegrationTest.java
@@ -1,19 +1,20 @@
 package cucumbertests;
 
-import com.springboothackathon.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootContextLoader;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ContextConfiguration;
+
+import com.springboothackathon.SpringBootApplication;
+
 
 @ContextConfiguration(
         classes = SpringBootApplication.class,
         loader = SpringBootContextLoader.class)
 @SpringBootTest(classes = SpringBootApplication.class,
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class SpringIntegrationTest {
 
 
 }
-
