+++ src/test/java/cucumbertests/stepdefs/CommonSteps.java
@@ -2,16 +2,14 @@ package cucumbertests.stepdefs;
 
 import com.springboothackathon.SpringBootApplication;
 import cucumber.api.java.en.Given;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.junit.Before;
 import org.springframework.boot.SpringApplication;
 
-public class CommonGivenSteps {
-    @Autowired
-    AppHttpClient appHttpClient;
+public class CommonSteps {
 
-    /*@Given("^the server is running$")
-    public void the_server_is_running()  {
+    @Before
+    //@Given("^the server is running$")
+    public void runServer()  {
         SpringApplication.run(SpringBootApplication.class);
-    }*/
-
+    }
 }
