--- src/test/java/cucumbertests/stepdefs/MovieControllerStepDefs.java
@@ -1,29 +1,0 @@
-package cucumbertests.stepdefs;
-
-import cucumber.api.PendingException;
-import cucumber.api.java.en.Given;
-import cucumber.api.java.en.Then;
-import cucumber.api.java.en.When;
-import cucumbertests.SpringIntegrationTest;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringApplication;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-
-public class MovieControllerStepDefs extends SpringIntegrationTest {
-
-    @Autowired
-    AppHttpClient appHttpClient;
-
-    StringBuffer response;
-
-    @Then("^client receives a list of all movies$")
-    public void client_receives_a_list_of_all_movies() throws Throwable {
-        // Write code here that turns the phrase above into concrete actions
-        throw new PendingException();
-    }
-}
